[FIX/#218] userId를 가지고 유저 정보를 가져올 때를 AuthRepo 를 사용하는 것은 적합하지 않으므로 UserRepo 를 따로 만들어서 데이터 관리하도록 수정
