+++ app/src/main/java/com/spoony/spoony/data/datasource/AuthRemoteDataSource.kt
@@ -6,6 +6,5 @@ import com.spoony.spoony.data.dto.response.UserSpoonCountResponseDto
 
 interface AuthRemoteDataSource {
     suspend fun getUserInfo(): BaseResponse<UserInfoResponseDto>
-    suspend fun getUserInfoById(userId: Int): BaseResponse<UserInfoResponseDto>
     suspend fun getSpoonCount(): BaseResponse<UserSpoonCountResponseDto>
 }
