+++ app/src/main/java/com/spoony/spoony/data/datasource/UserRemoteDataSource.kt
@@ -1,0 +1,8 @@
+package com.spoony.spoony.data.datasource
+
+import com.spoony.spoony.core.network.BaseResponse
+import com.spoony.spoony.data.dto.response.UserInfoResponseDto
+
+interface UserRemoteDataSource {
+    suspend fun getUserInfoById(userId: Int): BaseResponse<UserInfoResponseDto>
+}
