+++ app/src/main/java/com/spoony/spoony/data/di/DataSourceModule.kt
@@ -7,6 +7,7 @@ import com.spoony.spoony.data.datasource.MapRemoteDataSource
 import com.spoony.spoony.data.datasource.PlaceDataSource
 import com.spoony.spoony.data.datasource.PostRemoteDataSource
 import com.spoony.spoony.data.datasource.ReportDataSource
+import com.spoony.spoony.data.datasource.UserRemoteDataSource
 import com.spoony.spoony.data.datasourceimpl.AuthRemoteDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.CategoryDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.ExploreRemoteDataSourceImpl
@@ -14,6 +15,7 @@ import com.spoony.spoony.data.datasourceimpl.MapRemoteDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.PlaceDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.PostRemoteDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.ReportDataSourceImpl
+import com.spoony.spoony.data.datasourceimpl.UserRemoteDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -42,6 +44,12 @@ abstract class DataSourceModule {
     @Binds
     @Singleton
     abstract fun bindAuthRemoteDataSource(authRemoteDataSourceImpl: AuthRemoteDataSourceImpl): AuthRemoteDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindUserRemoteDataSource(
+        userRemoteDataSourceImpl: UserRemoteDataSourceImpl
+    ): UserRemoteDataSource
 
     @Binds
     @Singleton
