+++ app/src/main/java/com/spoony/spoony/data/di/RepositoryModule.kt
@@ -7,6 +7,7 @@ import com.spoony.spoony.data.repositoryimpl.MapRepositoryImpl
 import com.spoony.spoony.data.repositoryimpl.PostRepositoryImpl
 import com.spoony.spoony.data.repositoryimpl.RegisterRepositoryImpl
 import com.spoony.spoony.data.repositoryimpl.ReportRepositoryImpl
+import com.spoony.spoony.data.repositoryimpl.UserRepositoryImpl
 import com.spoony.spoony.domain.repository.AuthRepository
 import com.spoony.spoony.domain.repository.CategoryRepository
 import com.spoony.spoony.domain.repository.ExploreRepository
@@ -14,6 +15,7 @@ import com.spoony.spoony.domain.repository.MapRepository
 import com.spoony.spoony.domain.repository.PostRepository
 import com.spoony.spoony.domain.repository.RegisterRepository
 import com.spoony.spoony.domain.repository.ReportRepository
+import com.spoony.spoony.domain.repository.UserRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -50,4 +52,8 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindReportRepository(reportRepositoryImpl: ReportRepositoryImpl): ReportRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindUserRepository(userRepositoryImpl: UserRepositoryImpl): UserRepository
 }
