+++ app/src/main/java/com/spoony/spoony/data/di/ServiceModule.kt
@@ -7,6 +7,7 @@ import com.spoony.spoony.data.service.MapService
 import com.spoony.spoony.data.service.PlaceService
 import com.spoony.spoony.data.service.PostService
 import com.spoony.spoony.data.service.ReportService
+import com.spoony.spoony.data.service.UserService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -51,4 +52,9 @@ object ServiceModule {
     @Singleton
     fun provideMapService(retrofit: Retrofit): MapService =
         retrofit.create(MapService::class.java)
+
+    @Provides
+    @Singleton
+    fun provideUserService(retrofit: Retrofit): UserService =
+        retrofit.create(UserService::class.java)
 }
