+++ app/src/main/java/com/spoony/spoony/data/service/UserService.kt
@@ -1,0 +1,13 @@
+package com.spoony.spoony.data.service
+
+import com.spoony.spoony.core.network.BaseResponse
+import com.spoony.spoony.data.dto.response.UserInfoResponseDto
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface UserService {
+    @GET("/api/v1/user/{userId}")
+    suspend fun getUserInfoById(
+        @Path("userId") userId: Int
+    ): BaseResponse<UserInfoResponseDto>
+}
