+++ app/src/main/java/com/spoony/spoony/domain/repository/AuthRepository.kt
@@ -4,6 +4,5 @@ import com.spoony.spoony.domain.entity.UserEntity
 
 interface AuthRepository {
     suspend fun getUserInfo(): Result<UserEntity>
-    suspend fun getUserInfoById(userId: Int): Result<UserEntity>
     suspend fun getSpoonCount(): Result<Int>
 }
