+++ app/src/main/java/com/spoony/spoony/domain/repository/UserRepository.kt
@@ -1,0 +1,7 @@
+package com.spoony.spoony.domain.repository
+
+import com.spoony.spoony.domain.entity.UserEntity
+
+interface UserRepository {
+    suspend fun getUserInfoById(userId: Int): Result<UserEntity>
+}
